More olepres, losts of various fixes
