+++ extract_msg/enums.py
@@ -22,6 +22,7 @@ __all__ = [
     'ContactAddressIndex',
     'ContactLinkState',
     'DeencapType',
+    'DevModeFields',
     'DirectoryEntryType',
     'DisplayType',
     'DMPaperSize',
@@ -457,6 +458,29 @@ class DeencapType(enum.IntEnum):
     """
     PLAIN = 0
     HTML = 1
+
+
+
+class DevModeFields(enum.IntFlag):
+    DM_NUP           = 0b00000000000000000000000000000010
+    DM_SCALE         = 0b00000000000000000000000000001000
+    DM_PAPERWIDTH    = 0b00000000000000000000000000010000
+    DM_PAPERLENGTH   = 0b00000000000000000000000000100000
+    DM_PAPERSIZE     = 0b00000000000000000000000001000000
+    DM_ORIENTATION   = 0b00000000000000000000000010000000
+    DM_COLLATE       = 0b00000000000000000000000100000000
+    DM_TTOPTION      = 0b00000000000000000000001000000000
+    DM_YRESOLUTION   = 0b00000000000000000000010000000000
+    DM_DUPLEX        = 0b00000000000000000000100000000000
+    DM_COLOR         = 0b00000000000000000001000000000000
+    DM_PRINTQUALITY  = 0b00000000000000000010000000000000
+    DM_DEFAULTSOURCE = 0b00000000000000000100000000000000
+    DM_COPIES        = 0b00000000000000001000000000000000
+    DM_ICMMETHOD     = 0b00000000000000010000000000000000
+    DM_FORMNAME      = 0b00000000100000000000000000000000
+    DM_DITHERTYPE    = 0b00100000000000000000000000000000
+    DM_MEDIATYPE     = 0b01000000000000000000000000000000
+    DM_ICMINTENT     = 0b10000000000000000000000000000000
