+++ extract_msg_tests/constants.py
@@ -10,8 +10,10 @@ from pathlib import Path
 
 
 TEST_FILE_DIR = Path(__file__).parent.parent / 'example-msg-files'
-USER_TEST_DIR = None
+_utd = None
 if bool(userTestDir := os.environ.get('EXTRACT_MSG_TEST_DIR')):
     userTestDir = Path(userTestDir)
     if userTestDir.exists():
-        USER_TEST_DIR = userTestDir # type: ignore
+        _utd = userTestDir
+
+USER_TEST_DIR = _utd
